Streamline code for future functionality

+++ util/weatherapi.go
@@ -17,8 +17,8 @@ type WeatherAPI struct {
 }
 
 type Weather_Coord struct {
-	Lon float32 `json:"lon"`
-	Lat float32 `json:"lat"`
+	Lon float64 `json:"lon"`
+	Lat float64 `json:"lat"`
 }
 
 type Weather_Weather struct {
@@ -33,10 +33,10 @@ type Weather_Base struct {
 }
 
 type Weather_Main struct {
-	Temp float32 `json:"temp"`
-	Feel float32 `json:"feels_like"`
-	TMin float32 `json:"temp_min"`
-	TMax float32 `json:"temp_max"`
+	Temp float64 `json:"temp"`
+	Feel float64 `json:"feels_like"`
+	TMin float64 `json:"temp_min"`
+	TMax float64 `json:"temp_max"`
 	Pres int     `json:"pressure"`
 	Humi int     `json:"humidity"`
 }
@@ -46,7 +46,7 @@ type Weather_Visibility struct {
 }
 
 type Weather_Wind struct {
-	Spd float32 `json:"speed"`
+	Spd float64 `json:"speed"`
 	Deg int     `json:"deg"`
 }
 
@@ -61,7 +61,7 @@ type Weather_DT struct {
 type Weather_SYS struct {
 	Type int     `json:"type"`
 	ID   int     `json:"id"`
-	Mess float32 `json:"message"`
+	Mess float64 `json:"message"`
 	Ctry string  `json:"country"`
 	Sunr int     `json:"sunrise"`
 	Suns int     `json:"sunset"`
